cleanup code, move files around
