+++ src/components/screens/transactionDetails/components/index.js
@@ -3,7 +3,9 @@ import {
   Date, Amount, Fee, RequiredSignatures, Nonce,
 } from './components';
 
+import TransactionVotes from './transactionVotes';
+
 export {
   TransactionId, Sender, Recipient, Message, Illustration, Confirmations,
-  Date, Amount, Fee, RequiredSignatures, Nonce,
+  Date, Amount, Fee, RequiredSignatures, Nonce, TransactionVotes,
 };
