+++ src/components/screens/transactionDetails/components/styles.css
@@ -83,9 +83,21 @@
   grid-area: nonce;
 }
 
+.votes {
+  grid-area: votes;
+}
+
+.voteTag {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: flex-start;
+}
+
 .accountInfo {
   display: flex;
 }
+
 
 .addressRow {
   display: flex;
